Adds stateless/stateful registries
A stateless registry is usually used in conjunction with spectatord and
does not require to keep state on the current process. Tests done by
users of this library are greatly simplified if the registry and meters
can be queried.

This change adds a way to compose these behaviors to make testing easier
without having to add local state to meters when using the production
configuration.